Shield changes + future cleaving setup
Added shield attributes, these still need to be implemented into the knockback and damage methods in which they affect.  Also added a lang file.  Cleaving enchantment has been set up but there is no net.minecraft.util.Identifier in the external librarires so it may not be possible to add it.